Did You Like Me
Am a lady of 33 Years and a woman that has seen Live, And who know what it takes to have a good Man, and a marriage material , Yes am not the best, But i will like to be the best. So am looking for that good man. Know matter what people called this site. Msg me at d j u l i e 7 1 1 without spaces at gm. My ID is my email,  So go to my ID and email me with it. is my email, Its gmail ok